refactor: share token cache among all http clients
